Completely handle collections through Cursors (lazy-load, improved perf)

+++ src/org/tomahawk/libtomahawk/resolver/models/ScriptResolverTrack.java
@@ -33,7 +33,7 @@ public class ScriptResolverTrack {
 
     public String url;
 
-    public int duration;
+    public float duration;
 
     public String linkUrl;
